Refactor local image file functions to be DRY
