+++ okta/app.go
@@ -2,15 +2,10 @@ package okta
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"log"
 	"net/http"
-	"os"
 	"strings"
 	"sync"
 	"time"
@@ -92,7 +87,7 @@ var (
 			DiffSuppressFunc: func(k, old, new string, d *schema.ResourceData) bool {
 				return new == ""
 			},
-			StateFunc: logoStateFunc,
+			StateFunc: localFileStateFunc,
 		},
 		"logo_url": {
 			Type:        schema.TypeString,
@@ -723,25 +718,4 @@ func setAppUsersIDsAndGroupsIDs(ctx context.Context, d *schema.ResourceData, cli
 		_ = d.Set("users", convertStringSliceToSet(usersIDs))
 	}
 	return nil
-}
-
-func computeFileHash(filename string) string {
-	file, err := os.Open(filename)
-	if err != nil {
-		return ""
-	}
-	h := sha256.New()
-	if _, err := io.Copy(h, file); err != nil {
-		log.Fatal(err)
-	}
-	_ = file.Close()
-	return hex.EncodeToString(h.Sum(nil))
-}
-
-func logoStateFunc(val interface{}) string {
-	logoPath := val.(string)
-	if logoPath == "" {
-		return ""
-	}
-	return computeFileHash(logoPath)
 }
